+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * Upgrade to `hosc 0.12` and `hsc3 0.12`
 * Use `Sound.SC3.Server.Enum` data types for NRT file format
 * Rename `outputHeaderFormat` to `outputSoundFileFormat`
+* Add `runNRT` function and change type of `withNRT` and `NRTOptions`
 
 ### v0.7.0